SQL Programming Project
Implemented a very rudimentary database engine that is loosely based on a hybrid between MySQL and SQLite which operates entirely from the command line and API calls. • Created a database host application that interfaces with a backend SQL database to implement a DBMS for a Library Circulation Desk. This application uses a GUI interface developed using Swing (Java).